Release 1.1.2 :gem:
This is considered a bugfix release because the internal working of
stylesheets should be rearranged, not their external API. If this breaks
something, it might not have been used as intended.

This should also fix a bug I'm seeing where the first build of a site
doesn't properly build the style.scss file.