Add related posts to posts layout below post content.
It requires that the `site.include_related_posts` variable be set to
true in the `_config.yml` file. This should prevent unexpected changes
for those using the theme already and it's not something I would
consider 'minimal' behavior.

Don't forget to include the 'classifier-reborn' gem for Latent Symantic
Indexing. See the example's Gemfile (`example/Gemfile`) for an example.